Swap error checking to test www.isError instead of www.error
This is result of noticing the HoloLens was show Debug error logs for
some Spark calls, despite them working fine on the Standalone builds,
the Vive, in the editor and even in the HoloLens emulator. A bit of
digging found that www.error was showing non-null even when a valid
result was given with a code of 200. The isError flag appears to give
the expected behaviour, so switched all error checks to use that.